Add `bindHost` argument
Remove the `{grpc,rest,supervisor}Host` arguments and replace with
`host` and `bindHost`. The `host` argument is the hostname at which the
GCF emulator can be reached by clients, while the `bindHost` is the
address to which the emulator will bind its listener(s). This allows
the emulator to be running on a different host from the client, and
supports running the emulator behind a NAT (for example when running
as a docker container).